avance

+++ src/main/java/com/eci/cosw/springbootsecureapi/service/UserService.java
@@ -12,7 +12,7 @@ public interface UserService
 {
     List<User> getUsers();
 
-    User getUser( Long id );
+    User getUser( int id );
 
     User createUser( User user );
